+++ Freentship_Shipper/app.json
@@ -22,6 +22,12 @@
             "../src/assets/notification-sound-other.wav"
           ]
         }
+      ],
+      [
+          "expo-image-picker",
+        {
+          "photosPermission": "The app accesses your photos to let you share them with your friends."
+        }
       ]
     ],
     "updates": {
