+++ Freentship_Shipper/package.json
@@ -17,6 +17,7 @@
     "cosmiconfig": "^7.0.1",
     "expo": "~46.0.9",
     "expo-device": "^4.3.0",
+    "expo-image-picker": "~13.3.1",
     "expo-notifications": "~0.16.1",
     "expo-status-bar": "~1.4.0",
     "firebase": "9.6.11",
